+++ src/main/java/novamachina/novacore/NovaCore.java
@@ -2,9 +2,11 @@ package novamachina.novacore;
 
 import static novamachina.novacore.NovaCore.MOD_ID;
 
-import net.minecraftforge.fml.common.Mod;
+import net.neoforged.fml.common.Mod;
 
 @Mod(MOD_ID)
 public class NovaCore {
+  public NovaCore() {}
+
   public static final String MOD_ID = "novacore";
 }