+++ src/main/java/novamachina/novacore/bootstrap/NeoforgeBlockEntityTypeRegistry.java
@@ -1,14 +1,15 @@
 package novamachina.novacore.bootstrap;
 
+import net.minecraft.core.Registry;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.level.block.entity.BlockEntity;
-import net.minecraftforge.registries.ForgeRegistries;
 import novamachina.novacore.core.IRegistry;
 import novamachina.novacore.world.level.block.BlockEntityTypeDefinition;
 
-public class ForgeBlockEntityTypeRegistry
+public class NeoforgeBlockEntityTypeRegistry
     implements IRegistry<BlockEntityTypeDefinition<? extends BlockEntity>> {
   @Override
   public void register(BlockEntityTypeDefinition<? extends BlockEntity> entry) {
-    ForgeRegistries.BLOCK_ENTITY_TYPES.register(entry.getId(), entry.getType());
+    Registry.register(BuiltInRegistries.BLOCK_ENTITY_TYPE, entry.getId(), entry.getType());
   }
 }