+++ src/main/java/novamachina/novacore/bootstrap/NeoforgeCreativeModeTabRegistry.java
@@ -5,7 +5,7 @@ import net.minecraft.core.registries.BuiltInRegistries;
 import novamachina.novacore.core.IRegistry;
 import novamachina.novacore.world.item.CreativeModeTabDefinition;
 
-public class ForgeCreativeModeTabRegistry implements IRegistry<CreativeModeTabDefinition> {
+public class NeoforgeCreativeModeTabRegistry implements IRegistry<CreativeModeTabDefinition> {
   @Override
   public void register(CreativeModeTabDefinition entry) {
     Registry.register(BuiltInRegistries.CREATIVE_MODE_TAB, entry.id(), entry.tab());