+++ src/main/java/novamachina/novacore/bootstrap/NeoforgeLootModifierRegistry.java
@@ -1,16 +1,16 @@
 package novamachina.novacore.bootstrap;
 
-import net.minecraftforge.common.loot.LootModifier;
-import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraft.core.Registry;
+import net.neoforged.neoforge.common.loot.LootModifier;
+import net.neoforged.neoforge.registries.NeoForgeRegistries;
 import novamachina.novacore.common.loot.modifier.LootModifierDefinition;
 import novamachina.novacore.core.IRegistry;
 
-public class ForgeLootModifierRegistry
+public class NeoforgeLootModifierRegistry
     implements IRegistry<LootModifierDefinition<? extends LootModifier>> {
-  @Override
+
   public void register(LootModifierDefinition<? extends LootModifier> entry) {
-    ForgeRegistries.GLOBAL_LOOT_MODIFIER_SERIALIZERS
-        .get()
-        .register(entry.getId(), entry.getCodec());
+    Registry.register(
+        NeoForgeRegistries.GLOBAL_LOOT_MODIFIER_SERIALIZERS, entry.getId(), entry.getCodec());
   }
 }