+++ src/main/java/novamachina/novacore/bootstrap/NeoforgeStatRegistry.java
@@ -5,7 +5,7 @@ import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import novamachina.novacore.core.IRegistry;
 
-public class ForgeStatRegistry implements IRegistry<ResourceLocation> {
+public class NeoforgeStatRegistry implements IRegistry<ResourceLocation> {
   @Override
   public void register(ResourceLocation entry) {
     Registry.register(BuiltInRegistries.CUSTOM_STAT, entry, entry);