+++ src/main/java/novamachina/novacore/client/renderer/blockentity/AbstractBlockEntityRenderer.java
@@ -1,6 +1,6 @@
 package novamachina.novacore.client.renderer.blockentity;
 
-import javax.annotation.Nonnull;
+import net.minecraft.client.renderer.blockentity.BlockEntityRenderer;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Block;
@@ -8,18 +8,17 @@ import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 
-public abstract class BlockEntityRenderer<T extends BlockEntity>
-    implements net.minecraft.client.renderer.blockentity.BlockEntityRenderer<T> {
+public abstract class AbstractBlockEntityRenderer<T extends BlockEntity>
+    implements BlockEntityRenderer<T> {
 
-  protected BlockEntityRenderer() {
+  protected AbstractBlockEntityRenderer() {
     super();
   }
 
   // Added from ExCompressum, thanks to Blay09 for that piece of code :D
-  @Nonnull
-  public static BlockState getStateFromItemStack(@Nonnull final ItemStack itemStack) {
+  public static BlockState getStateFromItemStack(ItemStack itemStack) {
     if (itemStack.getItem() instanceof BlockItem blockItem) {
-      @Nonnull final Block block = blockItem.getBlock();
+      Block block = blockItem.getBlock();
       try {
         return block.defaultBlockState();
       } catch (Exception e) {