+++ src/main/java/novamachina/novacore/world/item/ItemDefinition.java
@@ -1,18 +1,16 @@
 package novamachina.novacore.world.item;
 
 import java.util.Objects;
-import lombok.Getter;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.ItemLike;
-import org.checkerframework.checker.nullness.qual.NonNull;
 
 public class ItemDefinition<T extends Item> implements ItemLike {
-  @Getter private final ResourceLocation id;
-  @Getter private final String englishName;
+  private final ResourceLocation id;
+  private final String englishName;
   private final T item;
-  @Getter private final ItemType type;
+  private final ItemType type;
 
   public ItemDefinition(String englishName, ResourceLocation id, T item, ItemType type) {
     Objects.requireNonNull(id, "id");
@@ -23,7 +21,6 @@ public class ItemDefinition<T extends Item> implements ItemLike {
   }
 
   @Override
-  @NonNull
   public T asItem() {
     return item;
   }
@@ -34,6 +31,18 @@ public class ItemDefinition<T extends Item> implements ItemLike {
 
   public ItemStack itemStack(int stacksize) {
     return new ItemStack(item, stacksize);
+  }
+
+  public ResourceLocation getId() {
+    return this.id;
+  }
+
+  public String getEnglishName() {
+    return this.englishName;
+  }
+
+  public ItemType getType() {
+    return this.type;
   }
 
   public enum ItemType {